% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_handlers.R
\name{alert_event_handler}
\alias{alert_event_handler}
\title{Handler for displaying node data as popup}
\usage{
alert_event_handler(warn_deprecated = FALSE)
}
\description{
Mainly useful for testing - Basically chucks the data for that particular node
into a popup.
}
\seealso{
Other js_handlers: 
\code{\link{embed_event_handler}()}
}
\concept{js_handlers}
