% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_graph.R
\name{attach_dependencies}
\alias{attach_dependencies}
\title{Attach Html Dependencies}
\usage{
attach_dependencies(graph, standalone = T)
}
\arguments{
\item{graph}{a graph made by \code{\link[drake:render_graph]{drake::render_graph()}}.}

\item{standalone}{logical, whether the current graph is rendered on
its own page, or is part of a larger rmarkdown document.}
}
\description{
This should be done for self-contained instances of visNetwork, when
not being rendered as part of a larger document.
It sets up some stylesheet + javascript dependencies.
}
\details{
Use it like:\preformatted{graph <- visNetwork::visNetwork()
graph \%<>\% attach_dependencies()
graph \%>\% visNetwork::visSave()
}
}
\concept{widget_dependencies}
