% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy-roclet-col.R
\name{col_roclet}
\alias{col_roclet}
\alias{roclet_process.roclet_col}
\alias{roclet_output.roclet_col}
\title{Roclet for processing the \verb{@col} tag}
\usage{
col_roclet()

\method{roclet_process}{roclet_col}(roc, blocks, env, base_path)

\method{roclet_output}{roclet_col}(roc, results, base_path, ...)
}
\description{
Roclet for processing the \verb{@col} tag
}
\concept{roxygen_roclet}
