% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_graph.R
\name{decorate_plan}
\alias{decorate_plan}
\title{Decorate Plan with a Rich html Description}
\usage{
decorate_plan(
  plan,
  cache,
  group = NULL,
  clusters = NULL,
  desc_colname = "desc",
  colname_out = desc_colname,
  lookup_cache = NULL,
  json_render = TRUE,
  ...
)
}
\arguments{
\item{group}{Optional character scalar, name of the column used to
group nodes into columns. All the columns names of your original \code{drake}
plan are choices. The other choices (such as \code{"status"}) are column names
in the \code{nodes} . To group nodes into clusters in the graph,
you must also supply the \code{clusters} argument.}

\item{clusters}{Optional character vector of values to cluster on.
These values must be elements of the column of the \code{nodes} data frame
that you specify in the \code{group} argument to \code{drake_graph_info()}.}

\item{desc_colname}{the name of the column that provides the markdown description}

\item{colname_out}{the name of the column in which to store the enriched description}

\item{lookup_cache}{the \code{storr::storr} object, generated by
\code{load_package_colspec()}, containing keys (given by column aliases)
mapping to column metadata lists.}

\item{json_render}{logical. if \code{TRUE}, render the output as json, otherwise, render it as \code{html}}
}
\description{
Decorate Plan with a Rich html Description
}
\seealso{
Other graph_decoration: 
\code{\link{extract_column_names}()},
\code{\link{link_col2doc}()}
}
\concept{graph_decoration}
