% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_handlers.R
\name{embed_event_handler}
\alias{embed_event_handler}
\title{Handler for embedding data into the legend}
\usage{
embed_event_handler()
}
\description{
If attached as onSelect behaviour for a visNetwork graph,
will allow clicking of nodes to replace the legend for the graph
with metadata about that node.
}
\details{
Presently requires JQuery to operate, so may not work when launched from
R session, but does work in pkgdown.

It looks like this:

function(props) {
container = this.body.container;
legend = $(container).parent().children("\link{id*='legendhtmlwidget'}")\link{0};
node = this.body.data.nodes.get(props.nodes\link{0});
$(legend).css({
'font-size': '10px',
'width' : '30\%',
'overflow-y' : 'scroll'
});
$(container).css('width', '70\%');
legend.innerHTML = node.on_select_col;
}
}
\seealso{
Other js_handlers: 
\code{\link{alert_event_handler}()}
}
\concept{js_handlers}
