% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_colnames.R
\name{extract_column_names}
\alias{extract_column_names}
\title{Extract Column Names for targets in Drake Plan}
\usage{
extract_column_names(
  plan,
  cache,
  group = NULL,
  clusters = NULL,
  colname_out = "target_column_list",
  ...
)
}
\arguments{
\item{group}{Optional character scalar, name of the column used to
group nodes into columns. All the columns names of your original \code{drake}
plan are choices. The other choices (such as \code{"status"}) are column names
in the \code{nodes} . To group nodes into clusters in the graph,
you must also supply the \code{clusters} argument.}

\item{clusters}{Optional character vector of values to cluster on.
These values must be elements of the column of the \code{nodes} data frame
that you specify in the \code{group} argument to \code{drake_graph_info()}.}

\item{colname_out}{the name given to the list-column containing each target's columns}
}
\description{
Given a \code{\link[drake:drake_plan]{drake::drake_plan()}}, extract the column names of each target, and store it
in a new column.
}
\seealso{
Other graph_decoration: 
\code{\link{link_col2doc}()}
}
\concept{graph_decoration}
