% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy-tag-col.R
\name{extract_named_captures}
\alias{extract_named_captures}
\title{Extract named captures from gregexpr}
\usage{
extract_named_captures(string, match_object)
}
\description{
Extract named captures from gregexpr
}
