% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{fake_documentation}
\alias{fake_documentation}
\title{Fake Documentation}
\description{
This is documentation
}
\section{Input Columns:}{
\describe{
\item{first_col}{This is a good col
\itemize{\item col_1\item col_1st}
}
\item{second_col}{This is a bad col
\itemize{\item col_2\item col_2nd}
}
}
}
\section{Output Columns:}{
\describe{
\item{col_out_1}{This col is gnarly!
\itemize{\item co1}
}
\item{col_out_2}{This col is terrible
\itemize{\item co2}
}
}
}
