% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_graph.R
\name{link_col2doc}
\alias{link_col2doc}
\title{Link Column Names to their Documentation.}
\usage{
link_col2doc(target_column_list, lookup_cache)
}
\arguments{
\item{target_column_list}{a list of character vectors specifying column names}

\item{lookup_cache}{the \code{storr::storr} object, generated by
\code{load_package_colspec()}, containing keys (given by column aliases)
mapping to column metadata lists.}
}
\value{
a character vector, of the same length as the outer dimension of
the input list, each element being a html table describing each
column.
}
\description{
Given a list-column containing column-names for targets (As part of a
\code{drake::drake_plan()} having been processed by \code{extract_column_names()}),
cross reference these columns with the \code{lookup_cache}, and produce html tables
for each set of columns.
}
\seealso{
Other graph_decoration: 
\code{\link{decorate_plan}()},
\code{\link{extract_column_names}()}
}
\concept{graph_decoration}
