% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_graph.R
\name{link_dependencies}
\alias{link_dependencies}
\title{Link Html Dependencies}
\usage{
link_dependencies(graph)
}
\description{
Link Html Dependencies
}
\seealso{
Other widget_dependencies: 
\code{\link{attach_dependencies}()}
}
\concept{widget_dependencies}
