% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{load_colspec_file}
\alias{load_colspec_file}
\title{Load colspec from a single file, to be imported into storr cache}
\usage{
load_colspec_file(path)
}
\description{
Load colspec from a single file, to be imported into storr cache
}
