% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{load_package_colspec}
\alias{load_package_colspec}
\title{Load a package colspec to case}
\usage{
load_package_colspec(pkg_name, lookup_cache = NULL)
}
\arguments{
\item{pkg_name}{the name of the package from which the set of columns
should be loaded.}

\item{lookup_cache}{the \code{storr::storr} object, generated by
\code{load_package_colspec()}, containing keys (given by column aliases)
mapping to column metadata lists.}
}
\description{
The package colspec may be consumed by \code{\link[=link_col2doc]{link_col2doc()}}, in order to
link column names to their metadata.
If \code{lookup_cache} is \code{NULL}/empty, a new one will be made.
}
