% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-mtcars.R
\name{mtcars_dataset}
\alias{mtcars_dataset}
\title{mtcars Dataset}
\description{
These cars go vroom!
}
\seealso{
Other example_datasets: 
\code{\link{lm_object}}
}
\concept{example_datasets}
\section{Output Columns:}{
\describe{
\item{mpg}{Numeric.
\code{mpg} or "Miles Per Gallon" is how many miles you can drive on
a gallon of fuel.
\itemize{}
}
\item{cyl}{Integer
Number of cylinders
\itemize{}
}
\item{disp}{Integer.
I literally don't know what this means
\itemize{}
}
\item{hp}{Integer
Horsepower is a measure of the number of miniature horses this car
can fit inside itself, which run on a treadmill that make it go fast.
\itemize{}
}
\item{drat}{Numeric
This is a word you say when something doesn't go the way you hoped.
\strong{Drat!}
\itemize{}
}
\item{wt}{Numeric
Stands for "What the..." - An exclamation of surprise!

Or \strong{weight}, maybe.
\itemize{}
}
\item{qsec}{Numeric.
Number of seconds you need to line up in a queue in order to buy this car.
\itemize{}
}
\item{vs}{Logical, Does the car ever compete?
\itemize{}
}
\item{am}{Logical, is this car okay to use in the morning, or should you save it for
after midday?
\itemize{}
}
\item{gear}{Integer,

How many different gears transmissions does this car have.
\itemize{}
}
\item{carb}{Integer, Number of grams of carbohydrates contained in this car.
Typically pretty low, as cars are not particularly nutritious.
\itemize{}
}
}
}
