% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy-roclet-col.R
\name{output_paths}
\alias{output_paths}
\title{Output paths}
\description{
Mandrake can derive the output path for its column mappings
from a few places:
}
\section{Output Paths Configuration}{

\itemize{
\item \strong{by default}, for a package \code{pkgname}, \code{mandrake} will store its output
under \code{"pkgname.yaml"} at:
\itemize{
\item pkgname/inst/mandrake
}
\item If you need to customize this, use:
\itemize{
\item This can be done by setting the following in \code{DESCRIPTION}:
\itemize{
\item \code{Roxygen: list(mandrake_output = "path/from/pkg/root")}
}
\item See the documentation for \link[roxygen2:load_options]{roxygen2::roxy_meta_get} for more information..
}
}
}

