% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_graph.R
\name{attach_dependencies}
\alias{attach_dependencies}
\title{Attach Html Dependencies}
\usage{
attach_dependencies(graph)
}
\description{
Attach Html Dependencies
}
\seealso{
Other widget_dependencies: 
\code{\link{link_dependencies}()}
}
\concept{widget_dependencies}
