% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_handlers.R
\name{embed_event_handler}
\alias{embed_event_handler}
\title{Handler for embedding data into the legend}
\usage{
embed_event_handler()
}
\description{
If attached as onSelect behaviour for a visNetwork graph,
will allow clicking of nodes to replace the legend for the graph
with metadata about that node.
}
\details{
Presently requires JQuery to operate, so may not work when launched from
R session, but does work in pkgdown.
}
\seealso{
Other js_handlers: 
\code{\link{alert_event_handler}()}
}
\concept{js_handlers}
