% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_colnames.R
\name{link_col2doc}
\alias{link_col2doc}
\title{Link Column Names to their Documentation}
\usage{
link_col2doc(target_column_list, lookup_cache)
}
\arguments{
\item{target_column_list}{a list of character vectors specifying column names}

\item{lookup_cache}{todo:doc me}
}
\description{
Link Column Names to their Documentation
}
